Beta-v1.5.8 (add mitra_id cust | modif dashboard admin)

+++ app/Http/Controllers/CustomerController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Customer;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class CustomerController extends Controller
@@ -21,7 +22,8 @@ class CustomerController extends Controller
      */
     public function create()
     {
-        return view('admin.customer_create');
+        $users = User::where('role', 'mitra')->get();
+        return view('admin.customer_create', compact('users'));
     }
 
     /**
@@ -32,6 +34,7 @@ class CustomerController extends Controller
         $request->validate([
             'nama' => 'required|string|max:64|unique:customers',
             'hp' => 'nullable',
+            'mitra_id' => 'nullable|exists:user,id'
         ]);
 
         Customer::create($request->all());
@@ -52,7 +55,9 @@ class CustomerController extends Controller
      */
     public function edit(string $id)
     {
-        //
+        $customer = Customer::findOrFail($id);
+        $users = User::where('role', 'mitra')->get();
+        return view('admin.customer_edit', compact('users', 'customer'));
     }
 
     /**
@@ -60,7 +65,20 @@ class CustomerController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        //
+        // Validasi input
+        $request->validate([
+            'nama' => 'required|string|max:64|unique:customers,nama,' . $id, // Mengabaikan ID yang sedang diupdate
+            'hp' => 'nullable',
+            'mitra_id' => 'nullable|exists:users,id' // Pastikan mitra_id ada di tabel users
+        ]);
+
+        // Temukan customer berdasarkan ID
+        $customer = Customer::findOrFail($id);
+
+        // Update data customer
+        $customer->update($request->all());
+
+        return redirect()->route('customer.index')->with('success', 'Customer berhasil diperbarui.');
     }
 
     /**
@@ -68,7 +86,11 @@ class CustomerController extends Controller
      */
     public function destroy(string $id)
     {
+        $user = User::get();
         $customers = Customer::findOrFail($id);
+        if ($user->role != 'superadmin') {
+            return redirect()->back()->with('error', 'Hanya superadmin yang dapat menghapus data ini');
+        }
         $customers->delete();
 
         return redirect()->route('customer.index')->with('success', 'Customer berhasil dihapus');
